+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Changed `cargo tarpaulin` command to use `--engine llvm` in GitHub Actions.
 - Changed `impl Chksum` for `PathBuf` - call method and DRY.
+- Changed `impl Chksum` for `File` - call method and DRY.
 
 ## [0.2.0] - 2023-08-12